grivances/requirement.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cryptography==40.0.2
defusedxml==0.7.1
Django==4.2
django-admin-interface==0.25.0
django-allauth==0.54.0
django-colorfield==0.8.0
django-crispy-forms==2.0
django-extensions==3.2.1
django-phonenumber-field==7.0.2
django-rest-framework==0.1.0
django-suit @ https://github.com/darklow/django-suit/tarball/v2#sha256=b40c916de9528ca86bcccd42d1fd54e7773e5071d2bfd9efc576039c1a441c0b
djangorestframework==3.14.0
idna==3.4
lazy-object-proxy==1.9.0
oauthlib==3.2.2
phonenumberslite==8.13.10
Pillow==9.5.0
psycopg2-binary==2.9.6
pycparser==2.21
PyJWT==2.6.0
pyotp==2.8.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
sqlparse==0.4.4
text-unidecode==1.3
toml==0.10.2
tzdata==2023.3
urllib3==1.26.15
wrapt==1.15.0
xlwt==1.3.0